Move resolver functions to `utils` pkg
Without it, the functions can't be imported since
Go identifies it as: "a program, not an
importable package"
